% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{adjust_for_inflation}
\alias{adjust_for_inflation}
\title{Adjust for inflation}
\usage{
adjust_for_inflation(value, date_from, date_to, base_year = 2015)
}
\arguments{
\item{value}{The \emph{nominal} value to be adjusted for inflation (in TND)}

\item{date_from}{The date of measurement # Improve}

\item{date_to}{The Target Date}

\item{base_year}{The base year used to calculate the underlying Consumer Price Index}
}
\value{
The \emph{real} value adjusted for inflation
}
\description{
Adjust for inflation
}
\examples{

adjust_for_inflation(
  value = 100,
  date_from = as.Date("2020-01-01"),
  date_to = as.Date("2022-01-01"),
  base_year = 2015
)
}
