% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{adjust_for_inflation2}
\alias{adjust_for_inflation2}
\title{Adjust for inflation (For dataframes)}
\usage{
adjust_for_inflation2(
  data,
  values_col,
  date_from_col,
  date_to_col,
  base_year_col,
  output_col = "real_value"
)
}
\arguments{
\item{data}{dataframe}

\item{values_col}{column name containing amounts to be adjusted for inflation}

\item{date_from_col}{column name containing the observation dates}

\item{date_to_col}{column name containing the target dates for adjustment}

\item{base_year_col}{column name containing the base year # TODO: make it optional}

\item{output_col}{column name for output values}
}
\value{
dataframe with added column for real values
}
\description{
Adjust for inflation (For dataframes)
}
\examples{
\dontrun{
data <- tibble::tibble(
  budget = c(1, 1, 1, 1),
  observation_date = c(
    as.Date("2016-01-01"),
    as.Date("2017-01-01"),
    as.Date("2018-01-01"),
    as.Date("2019-01-01")
  ),
  target_date = c(
    as.Date("2015-01-01"),
    as.Date("2015-01-01"),
    as.Date("2015-01-01"),
    as.Date("2015-01-01")
  ),
  base_year = c(2015, 2015, 2015, 2015)
)
data \%>\%
  adjust_for_inflation2(
    values_col = budget,
    date_from_col = observation_date,
    date_to_col = target_date,
    base_year_col = base_year
  )
}
}
