% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{calculate_inflation}
\alias{calculate_inflation}
\title{Calculate Inflation}
\usage{
calculate_inflation(date_from, date_to, base_year = 2015)
}
\arguments{
\item{date_from}{Initial date}

\item{date_to}{Target date}

\item{base_year}{The base year used to calculate the underlying Consumer Price Index}
}
\value{
Inflation rate between the two provided dates.
}
\description{
Calculate Inflation
}
\examples{

calculate_inflation(
  date_from = as.Date("2020-01-01"),
  date_to = as.Date("2022-01-01"),
  base_year = 2015
)
}
