% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consumer_price_index}
\alias{consumer_price_index}
\title{Consumer Price Index.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{base_year}{Base year used for measurements. This is important as a change in base year reflects a change in the consumer basket used.}
\item{cpi}{The value of the Consumer Price Index}
\item{measurement_date}{The date of measurement (Month)}
...
}
}
\source{
\url{"http://www.ins.tn/publication/indice-des-prix-la-consommation"}
}
\usage{
consumer_price_index
}
\description{
A dataset containing the consumer price index as measured by the
Tunisian Statistics' National Institute (INS) since 1962.
}
\keyword{datasets}
