% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ins_coicop}
\alias{ins_coicop}
\title{National Institute of Statistics (INS) to COICOP correspondence table.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{ins_category}{The category name used by the INS}
\item{ins_weight_2015}{The weight of the category in the base basket of goods for 2015}
\item{id_coicop}{The COICOP ID}
\item{level}{The level of the category in the hierarchy}
\item{division}{The COICOP division}
\item{group}{The COICOP group}
\item{class}{The COICOP class}
\item{coicop_category_fr}{The COICOP category name in French}
}
}
\source{
\url{"http://www.ins.tn/publication/indice-des-prix-la-consommation-juillet-2022"}
}
\usage{
ins_coicop
}
\description{
A correspondence table between the categories used by the INS for the
consumer price index and the international standard COICOP.
The dataset includes also the weights used by the INS for the base year 2015.
}
\keyword{datasets}
