% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_inflation}
\alias{item_inflation}
\title{Inflation (monthly and annual) by item since 2015, using 2015 as a base year.}
\format{
A data frame with 6097 rows and 6 variables:
\describe{
\item{ins_category}{The category name used by the INS}
\item{ins_weight_2015}{The weight of the category in the base basket of goods for 2015}
\item{measurement_period}{The month of measurement}
\item{cpi}{The value of the Consumer Price Index on the specified date}
\item{monthly_item_inflation}{Inflation of the item between the specified month and the previous one}
\item{yearly_item_inflation}{Inflation of the item between the specified month and the same month from a year before}
}
}
\source{
\url{"http://www.ins.tn/publication/indice-des-prix-la-consommation-juillet-2022"}
}
\usage{
item_inflation
}
\description{
Inflation (monthly and annual) by item since 2015, using 2015 as a base year.
}
\keyword{datasets}
