% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuninflation.R
\name{tuninflation}
\alias{tuninflation}
\title{tuninflation}
\usage{
tuninflation(
  base_year = 2015,
  start_date,
  end_date,
  by_item = FALSE,
  frequency = "yearly",
  relative_to = "year"
)
}
\arguments{
\item{base_year}{The base year for CPI values}

\item{start_date}{The date from which to start computing inflation}

\item{end_date}{The date at which to stop computing inflation}

\item{by_item}{Whether to compute inflation broken down by consumption components}

\item{frequency}{The frequency of reported results, can be either 'yearly' or 'monthly'}

\item{relative_to}{The reference point in time relative to which inflation is computed. Can be either 'year' or 'month'}
}
\value{
A dataframe with an index matching the provided frequency, and column for inflation computed relative to the provided frame of reference
}
\description{
tuninflation
}
